move test constants to shared folder
